+++ modules/security/classes/general/session_mc.php
@@ -4,6 +4,8 @@ class CSecuritySessionMC
 	/** @var Memcache $connection*/
 	protected static $connection = null;
 	protected static $sessionId = null;
+	protected static $isReadOnly = false;
+	protected static $isSessionReady = false;
 
 	/**
 	 * @return bool
@@ -12,6 +14,8 @@ class CSecuritySessionMC
 	{
 		if(self::isConnected())
 			return true;
+
+		self::$isReadOnly = defined('BX_SECURITY_SESSION_READONLY');
 
 		return self::newConnection();
 	}
@@ -26,17 +30,25 @@ class CSecuritySessionMC
 		return CSecuritySessionMC::Init();
 	}
 
+	/**
+	 * @return bool
+	 */
 	public static function close()
 	{
 		if(!self::isConnected() || !self::isValidId(self::$sessionId))
-			return;
+			return false;
 
-		if(isSessionExpired())
-			self::destroy(self::$sessionId);
+		if (!self::$isReadOnly && self::$isSessionReady)
+		{
+			if(isSessionExpired())
+				self::destroy(self::$sessionId);
 
-		self::$connection->delete(self::getPrefix().self::$sessionId.".lock");
+			self::$connection->delete(self::getPrefix().self::$sessionId.".lock");
+		}
+
 		self::$sessionId = null;
 		self::closeConnection();
+		return true;
 	}
 
 	/**
@@ -48,23 +60,43 @@ class CSecuritySessionMC
 		if(!self::isConnected() || !self::isValidId($id))
 			return "";
 
-		$lockTimeout = 55;//TODO: add setting
-		$lockWait = 59000000;//micro seconds = 60 seconds TODO: add setting
-		$waitStep = 100;
 		$sid = self::getPrefix();
 
-		while(!self::$connection->add($sid.$id.".lock", 1, 0, $lockTimeout))
+		if (!self::$isReadOnly)
 		{
-			usleep($waitStep);
-			$lockWait -= $waitStep;
-			if($lockWait < 0)
-				CSecuritySession::triggerFatalError('Unable to get session lock within 60 seconds.');
+			$lockTimeout = 55;//TODO: add setting
+			$lockWait = 59000000;//micro seconds = 60 seconds TODO: add setting
+			$waitStep = 100;
 
-			if($waitStep < 1000000)
-				$waitStep *= 2;
+			if (defined('BX_SECURITY_SESSION_MEMCACHE_EXLOCK') && BX_SECURITY_SESSION_MEMCACHE_EXLOCK)
+				$lock = Bitrix\Main\Context::getCurrent()->getRequest()->getRequestedPage();
+			else
+				$lock = 1;
+
+			while(!self::$connection->add($sid.$id.".lock", $lock, 0, $lockTimeout))
+			{
+				usleep($waitStep);
+				$lockWait -= $waitStep;
+				if($lockWait < 0)
+				{
+					$errorText = 'Unable to get session lock within 60 seconds.';
+					if ($lock !== 1)
+					{
+						$lockedUri = self::$connection->get($sid.$id.".lock");
+						if ($lockedUri && $lockedUri != 1)
+							$errorText .= sprintf(' Locked by "%s".', self::$connection->get($sid.$id.".lock"));
+					}
+
+					CSecuritySession::triggerFatalError($errorText);
+				}
+
+				if($waitStep < 1000000)
+					$waitStep *= 2;
+			}
 		}
 
 		self::$sessionId = $id;
+		self::$isSessionReady = true;
 		$res = self::$connection->get($sid.$id);
 		if($res === false)
 			$res = "";
@@ -74,55 +106,75 @@ class CSecuritySessionMC
 
 	/**
 	 * @param string $id - session id, must be valid hash
-	 * @param array $sessionData
+	 * @param string $sessionData
+	 * @return bool
 	 */
 	public static function write($id, $sessionData)
 	{
 		if(!self::isConnected() || !self::isValidId($id))
-			return;
+			return false;
+
+		if (!self::$isSessionReady)
+			return false;
+
+		if (self::$isReadOnly)
+			return true;
 
 		$sid = self::getPrefix();
-		$maxlifetime = intval(ini_get("session.gc_maxlifetime"));
+		$maxLifetime = intval(ini_get("session.gc_maxlifetime"));
 
 		if(CSecuritySession::isOldSessionIdExist())
-			$oldSessionId = CSecuritySession::getOldSessionId();
-		else
-			$oldSessionId = $id;
+		{
+			$oldSessionId = CSecuritySession::getOldSessionId(true);
+			self::$connection->delete($sid.$oldSessionId);
+		}
 
-		self::$connection->delete($sid.$oldSessionId);
-		self::$connection->set($sid.$id, $sessionData, 0, time() + $maxlifetime);
+		self::$connection->set($sid.$id, $sessionData, 0, $maxLifetime);
+
+		return true;
 	}
 
 	/**
 	 * @param string $id - session id, must be valid hash
+	 * @return bool
 	 */
 	public static function destroy($id)
 	{
 		if(!self::isValidId($id))
-			return;
+			return false;
+
+		if (!self::$isSessionReady)
+			return false;
+
+		if (self::$isReadOnly)
+			return false;
 
 		$isConnectionRestored = false;
 		if(!self::isConnected())
 			$isConnectionRestored = self::newConnection();
 
 		if(!self::isConnected())
-			return;
+			return false;
 
 		$sid = self::getPrefix();
 		self::$connection->delete($sid.$id);
 
 		if(CSecuritySession::isOldSessionIdExist())
-			self::$connection->delete($sid.CSecuritySession::getOldSessionId());
+			self::$connection->delete($sid.CSecuritySession::getOldSessionId(true));
 
 		if($isConnectionRestored)
 			self::closeConnection();
+
+		return true;
 	}
 
 	/**
 	 * @param int $maxLifeTime - unused on this handler
+	 * @return bool
 	 */
 	public static function gc($maxLifeTime)
 	{
+		return true;
 	}
 
 	/**
@@ -147,7 +199,11 @@ class CSecuritySessionMC
 	 */
 	protected static function isValidId($pId)
 	{
-		return (bool) preg_match("/^[\\da-z]{6,32}$/iD", $pId);
+		return (
+			$pId
+			&& is_string($pId)
+			&& preg_match('/^[\da-z\-,]{6,}$/iD', $pId)
+		);
 	}
 
 	/**
